backend/pyproject.toml
[project]
name = "easylife-auth"
version = "1.0.0"
description = "EasyLife Admin Panel - FastAPI Backend with Role-Based Access Control"
authors = [{ name = "EasyLife Team" }]
license = "LicenseRef-Proprietary"
requires-python = ">=3.12"
dependencies = [
    # FastAPI and server
    "fastapi>=0.110.0,<1.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
    "python-multipart>=0.0.12,<1.0.0",
    # Database - Motor for async MongoDB
    "motor>=3.3.0,<4.0.0",
    # Authentication
    "pyjwt>=2.9.0,<3.0.0",
    "werkzeug>=3.0.0,<4.0.0",
    # Email - async
    "aiosmtplib>=3.0.2,<6.0.0",
    # Validation
    "pydantic>=2.5.0,<3.0.0",
    "email-validator>=2.1.0,<3.0.0",
    # Utilities
    "python-dotenv>=1.0.0,<2.0.0",
    "psutil>=5.9.0,<8.0.0",
    # File processing
    "pandas>=2.2.0,<4.0.0",
    "openpyxl>=3.1.0,<4.0.0",
    # Password hashing
    "bcrypt>=4.1.0,<6.0.0",
]

[project.optional-dependencies]
gcs = ["google-cloud-storage>=2.14.0,<4.0.0"]
jira = ["jira>=3.10.5,<4.0.0"]
atlassian = ["atlassian-python-api>=3.41.0"]
all = [
    "google-cloud-storage>=2.14.0,<4.0.0",
    "jira>=3.10.5,<4.0.0",
    "atlassian-python-api>=3.41.0",
]

[tool.poetry]
packages = [{ include = "easylifeauth", from = "src" }]

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0,<10.0.0"
pytest-asyncio = ">=0.23.0,<2.0.0"
pytest-cov = ">=5.0.0,<8.0.0"
httpx = ">=0.26.0,<1.0.0"
aiohttp = ">=3.9.0,<4.0.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = "-v --cov=src --cov-report=term-missing --cov-report=html --cov-report=xml:coverage.xml --cov-fail-under=95"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]
pythonpath = [".", "src"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
